+++ diesel/src/query_builder/mod.rs
@@ -21,6 +21,7 @@ mod select_statement;
 pub mod where_clause;
 pub mod insert_statement;
 pub mod update_statement;
+mod union_query;
 
 pub use self::bind_collector::BindCollector;
 pub use self::query_id::QueryId;
@@ -30,6 +31,8 @@ pub use self::select_statement::{SelectStatement, BoxedSelectStatement};
 pub use self::update_statement::{IncompleteUpdateStatement, AsChangeset, Changeset, UpdateTarget};
 #[doc(inline)]
 pub use self::insert_statement::IncompleteInsertStatement;
+#[doc(inline)]
+pub use self::union_query::UnionQuery;
 
 use std::error::Error;
 
@@ -140,3 +143,6 @@ impl<T: Query> AsQuery for T {
         self
     }
 }
+
+#[doc(hidden)]
+pub trait CombinableQuery: Query {}